+++ cypress/component/Input.cy.jsx
@@ -1,11 +1,11 @@
 import Input from "@/components/dynamic/Input";
 import { useState } from "react";
-import { USER } from "../../src/data/dynamic/forms/User";
+import { HELPER } from "../../src/data/dynamic/forms/Participant";
 
 describe("Input", () => {
   it("Placeholder", () => {
     const Parent = () => {
-      const [user, setUser] = useState(USER);
+      const [user, setUser] = useState(HELPER);
       return (
         <Input
           name="first"
@@ -28,7 +28,7 @@ describe("Input", () => {
 
   it("Typing...", () => {
     const Parent = () => {
-      const [user, setUser] = useState(USER);
+      const [user, setUser] = useState(HELPER);
       return (
         <Input
           name="first"
@@ -51,7 +51,7 @@ describe("Input", () => {
 
   it("Backspace", () => {
     const Parent = () => {
-      const [user, setUser] = useState(USER);
+      const [user, setUser] = useState(HELPER);
       return (
         <Input
           name="first"
@@ -75,7 +75,7 @@ describe("Input", () => {
 
   it("Select all clear", () => {
     const Parent = () => {
-      const [user, setUser] = useState(USER);
+      const [user, setUser] = useState(HELPER);
       return (
         <Input
           name="first"
