+++ src/components/dynamic/forms/Mentor.jsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import Form from "@/components/dynamic/forms/Form.jsx";
-import { FIELDS } from "../../../data/dynamic/forms/Mentors.js";
+import { FIELDS, HELPER } from "../../../data/dynamic/forms/Mentors.js";
 import { useSession } from "next-auth/react";
 import axios from "axios";
 import toast from "react-hot-toast";
@@ -11,6 +11,7 @@ const Mentor = () => {
   const { data: session } = useSession();
 
   const [mentor, setMentor] = useState({
+    ...HELPER,
     name: session.user.name,
     email: session.user.email,
   });
