+++ src/data/dynamic/forms/Availability.js
@@ -1,12 +1,3 @@
-export const HELPER = {
-  phone: "",
-  major: "",
-  grade: "",
-  gender: "",
-  shirt: "",
-  discord: "",
-};
-
 export const AVAILABILITY = {
   fridayMorning: {
     state: false,
@@ -56,13 +47,4 @@ export const AVAILABILITY = {
     state: false,
     text: "Sunday Night",
   },
-};
-export const FEEDBACK = {
-  "What was most helpful about the event?": "",
-  "Was there anything that you would have liked to see?": "",
-  "Was there anything that you did not find beneficial about the hackathon?":
-    "",
-  "How would you rate the hackathon?": "",
-  "How did you find the event?": "",
-  "Is there anything else you'd like to let us know?": "",
 };
