--- src/data/dynamic/forms/Judges.js
@@ -1,10 +1,0 @@
-export const JUDGES = {
-  first: "",
-  last: "",
-  email: "",
-  phone: "",
-  shirt: "",
-  title: "",
-  major: "",
-  photo: null,
-};
