+++ src/data/dynamic/forms/Mentors.js
@@ -124,3 +124,16 @@ export const FIELDS = {
     ],
   },
 };
+
+export const HELPER = {
+  name: "",
+  email: "",
+  phone: "",
+  discord: "",
+  major: "",
+  grade: "",
+  availability: {},
+  gender: "",
+  shirt: "",
+  response: "",
+};
