--- src/data/dynamic/forms/User.js
@@ -1,13 +1,0 @@
-export const USER = {
-  first: "",
-  last: "",
-  age: "",
-  phone: "",
-  email: "",
-  school: "",
-  major: "",
-  grade: "",
-  gender: "",
-  shirt: "",
-  resume: "",
-};
